requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
celery==5.3.4
boto3==1.34.14
pandas==2.1.4
numpy==1.26.3
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
mypy==1.8.0
sentry-sdk==1.39.2
stripe==7.12.0
sendgrid==6.11.0
twilio==8.11.0
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
python-dotenv==1.0.0